Map edit:
Adds fire extinguishers around the station.
Removes the safes from the armory and captain office.
Adds a safe to the vault containing 7 instant concrete grenades.
Adds a contraband locker to the armory, also moves the DRAGnet and the HoS gun into this locker. Said locker has a small chance to contain teargas clusterbusters.
Slightly edits the supermatter engine chamber in order to better contain the heat it produces when cooling is incorrectly setup.
Moves the engineering firefighting closets to the secure storage in engineering.
Adds a lot of coins around the station.
Replaces a burn kit on the shuttle with a poison kit.
Disables AI/PDA control of the dangerous old medibot.
Adds an additional monitor decryption key paper to telecomms.