Improves performance of set_airlock_overlays
They will now only update when they actually change. On my testing, this
made performance improve about 3-15 times, depending on the use case.

I could have made this a loop if I turned the overlays into a list with
defines but I was lazy and this copy-paste is probably slightly faster
than a list as well (despite being harder to read and to maintain).