Coffeescript; Bower; Gulp; asset_cache
We now write frontend code in Coffeescript, alongside our LESS.
PostCSS adds hacks to make CSS work on old versions of IE.
Bower is used to manage dependencies.
Gulp is used to build the code.

Images are dropped in favor of SVG or CSS effects.

NanoUI is now a 'simple' asset in the asset_cache