formally defines apprentices as a kind of antag instead of as "traitors".
Wizard events that spare the wizard should now properly spare apprentices too.

Moves renaming as an apprentice to the end of the spawn proc because the input prompt stops various important things (including the sound meant to inform people they got apprentice) from happening until after it's dealt with.

Restores the "+" for positive quality items in RPG loot. The reason this was changed was already fixed, it just was mistakenly never put back.