Tweaks blob expansion to be faster near the expanding blob but slower near the edges of the blob Blob spores now produce a slightly larger cloud of smoke when dying. Fragile spores and blob zombies produce the current, smaller cloud.
Replicating Foam will try to expand when damaged much more often.
Adds Electromagnetic Web, which does burn damage and EMPs targets, as well as causing a small EMP when dying, but takes more damage; a normal blob that hasn't been near a node or the core for 6~ seconds can be oneshotted by a laser.