Lava Proofs some mechs.
The firefighter ripley and the maurader mecha family are now lava proof.

Firefighters now have a better niche as the only mech available to the crew that can cross lava.

The only version of the Maurader available to players is the Mauler, and on average, nuke ops and lava will never come in contact.