Oh-cee, the original content skeleton
:cl: coiax, Nienhaus
rscdel: Unfortunately, the Belowyarn toys have been withdrawn from use
due to Assistants choking on the batteries. Centcom has sent their
condolences to the battery manufacturing plant.
rscadd: We are introducing Oh-cee the original content skeleton figurines.
Pull its string, and hear it say a variety of phrases. Not suitable for
infants or assistants under 36 months.
tweak: Note that pulling the string of all talking toys is now visible
to others.
/:cl:

With thanks to Nienhaus for doing the sprite.

- The toy AI has also been made a "talking" toy because it works the
  same as the others.