Changes how hidden fingerprints (for admin use) work.
Old system:
If the player is not the last person to touch the object, we add a new line to the hiddenfingerprints list
with key, real name, whether the mob has gloves, and time stamp.

Issues: If only one player touches the object then you only know about the first time they touched it. It's
not completely spam proof, two players touching the object alternatively can make the list as big as they
want.

New system:
Each player touching an object is attributed one line in the hiddenfingerprints list. The line
shows the player key, real name, and time stamp for both first and last time touched, and whether wearing gloves for both.

Issues: The list can become relatively big if many different players touch the same object (but not infinite).

If multiple players touch the object alternatively, you get more info with the old system. If one player is
the only one touching the object and does so multiple times, then you get more info with the new system.