Clone pod fixes, changes, tweaks
:cl: coiax
rscadd: Clone pods now notify the medical channel on a successful clone.
They also notify the medical channel if the clone dies or is horribly
mutilated; but not if it is ejected early by authorised medical
personnel.
rscadd: Cloning pods now are more capable of cloning humanoid species
that do not breathe.
/:cl:

- Cleaned up cloning.dm, removed a bunch of unused code, use TRUE and
FALSES, removed trailing returns, put some more colourful messages in.
- Small amounts of Bicardine are added to the clone to cope with any
NOBREATH crit damage.
- The `grab_ghost_when` var has been added, which will change when we
actually grab the ghost and stick it in the new body. By default it is
unchanged to when the clone is freshly made, but can be changed to
"mature" which is when the clone is ejected.