The true ruination of space: Death to the donut.
All existing structures in space outside of mining/lavaland, the station, and centcom have been converted to ruins.

z10 has been added to maps with lavaland to make the number of crosslinked z levels up to a solid 9.

z10 is the "true deep space" level where nothing can ever spawn, this is mostly because it doesn't exist in some maps.