Refactors all (F) links into using a define
:cl:
bugfix: Clicking the (F) link when an AI talks in binary chat will
follow its camera eye, the same as when (F) is clicked for its radio
chat.
/:cl:

FOLLOW_LINK is a define that creates the clickable (F) link, so it
doesn't have to be typed painstakingly out every time we want to give
that link to an observer.