Gives beepsky a home #18615
Beepsky now has his own home in the labor shuttle room.

This replaces the n2o storage with a small home for beepsky. There's a small bedsheet for him to lay on, a letter from his mom, and an emergency battery. An extra n2o canister has been put in the execution room to compensate for the loss of the storage.