Ore redemption machine streamlining.  (#19062)
Makes the ore redemptor machine 100000% less spammy to request consoles, and prints a list of what exactly is in the machine. The new non-spamminess means chemistry and the bar can now get a notification, because there will only be one message per load.

Adds bluespace polycrystals, basically a "smelted" "sheet" of bluespace crystals, to the game, in order to make the ore redemptor accept them.