Electric Guitar soundrewrite and mapping #20818
It's still loud and ear hurting but now it's in tune and shorter. You're welcome. Made for Okand's birthday. Now it maps the abomination as well so people can finally shred without admin help to spawn. The range is now shorter with less notes but it plays all the songs with no issues. Sounds were generated with a digital audio workstation and modified by me a lot.