Lich phylacteries are now points of interest
:cl: coiax
add: Lich phylacteries are now in the "points of interest" for ghosts.
/:cl:

Also, a phylactery just appends to the description of the item, rather
than clobbering it.

Balance consideration: if a lich summons magic and a crewmember get a
scrying orb, they'll be able to locate the phylactery immediately. So I
guess this is a lich nerf?