Reverted F8 to be Invismin and F7 to be build-mode
Having F8 as a Invisimin toggle key is increadibly useful for events and the like (it has always been that key as well, so reverted to that key for legacy issues).
F7 will be build mode (because having it on a key is actually really useful).
Having "Stealthmin" on a hot key REALLY isnt necessary. Its not something you toggle on and off frequently. Its something you set and leave and honestly only a few admins utilise it anyway.

(Also my first PR woo!)