Traitor, a traitor is a traitor.
Means that one of the crew members works for a criminal Syndicate and
must something to seize / kill someone / someone to protect and dump
from that Durkee on the shuttle. At English stations the word everyday,
for Russians for the mention of the word "traitor" is not in the chat
(IC) admins are stroking the head with a bangammer. So this word and
memetic.