Applying PDA overhaul Upstream (#30613)
* Applying PDA overhaul Upstream

Yeet?

* Used lordpidey's change

Like this?

* The build failed so I'm changing it back

* Update PDA.dm

* Update PDA.dm

* Make the settings visible only on the main screen

mode 0 is the hub, and the hub is the main menu. I think it'd be a lot more convenient for players to always have these options on the main menu if they need them. Makes it really easy to swap colors and fonts at the beginning of the round. I can put it in a settings menu if everyone really thinks it's necessary, but then someone would have to help me actually make the damn icons show up.

* Update PDA.dm

* Update PDA.dm

* Wait, I'm an idiot. I was editing the wrong thing.

* Here's your switch statements boi

* Declare font_index at the "Toggle_Font" code

I declared it exactly where it's relevant, but I think this looks unclean. I also redefined it which I actually needed to do.

* Replaced it where it was

cuz the build broke

* Make the color changer a picker instead of a field

An attempt

* Made these defines

@optimumtact @Cyberboss Like this?