Changelings learn languages from people they absorb
:cl: coiax
add: A changeling will learn all languages from anyone they absorb.
/:cl:

Eat a lizard to be able to fully impersonate a lizard with Draconic and
all that. Eating the Curator DOES teach you all languages, but you'll
only be able to speak the ones that your current tongue allows you to.