Fixes experimentor pickweight and goo runtime (#37822)
The experimentor is less likely to produce food when it transforms an item

The experimentor no longer breaks when the toxic waste malfunction occurs multiple times and isn't cleaned up

The experimentor irridiate function actually chooses what item to make properly rather than picking from the start of the list


The food weight change is because the probweight var wasn't effective because food was added to the list before anything that would increase the var, it's so low because there's a ton of food and it should be possible to get other things. I tested it in game and food is still the most common thing to get, I could bump up the prob for non-food things if needed.

The toxic waste runtime would make the experimentor unfixable except for reconstructing, and that resets the experiments.

The previous method of picking was broken and pretty poorly coded, I moved it to an associative list and the existing and non-broken pickweight function (not sure if this is also a code change).