Corrects a stat error in the Nuke Ops uplink (#38980)
When this was buffed 2 years ago the person who did it forgot to update the
uplink descriptions, so I've done so. Not sure if a changelog would be required
for this as it's an incredibly minor change, but I've added it anyway. Also not
sure about using the grammar tag, so please tell me if it's wrong to use it for
this sort of thing.