Pirate buffs via changes to pirate_default.dmm (#39072)
* Pirate buffs via changes to pirate_default.dmm

It's a buff to the Pirate ship, so they have carbon dioxide jetpacks (2), additional X4 (3), a Cannister to refill said jetpacks (1), and an L6 SAW (1; and honestly the thing Im most worried about, balance-wise; default ammo). Also gives the captain a renamed bottle of rum in his room.

* Update pirate_default.dmm

Removed the extra X4, Gun, and CO2 canister due to feedback.