[READY] Incinerator clean up and update (#39299)
Main goal: clean up atmos so it is more user friendly and up to date with its current uses
Side goal: provide the backbone of making tritium for the user while still requiring several edits to make it at a rate that can be used.
Reasoning: atmos can take a lot of math and a lot of trivial work to make it reach its end teir developements, so much so that many, especially newer atmos techs arent able to make anything tangible until around the 30 minute mark which is the length of most rounds that arent on low pop hours. When it comes to more advanced incinerators users the filter, scrubber, and mixer are all mandatory changes for a tritium set up which can become very boring and just function as a time gate between the start of the round and tritium to work with. These changes are meant to remove the less thought requiring parts of incinerator changes and encourage newer players to make their first exotic gas (similiar to how each department has a simple entry level job that a player can get hooked on to start). The changing in scrubbers and distro are meant to allow players to make many common changes without removing parts the scrubbers and distro from the room itself and the removing of the old piping for the same reason. Removing the old piping will also allow the room to be less cluttered after the incinerator user makes his or her changes. Toxins incinerator was also updated for universality