Add zombies to box whiteship (#39643)
These, like in the other ships, are pretty weak (meta still has the strongest
monsters, but it's also the most well stocked ship). These zombies have no
infection chance, 60 hp instead of 100, and do 11 damage instead of 21 (could
be raised to 15, if you guys want). And like the other ships there are 4
monsters on it.