Drink quality affects mood (#39628)
Time to get fancy, bartenders. The harder it is to make a drink, the stronger
the effect, usually. There are exceptions, but quality tiers look something
like this:

  1. any mixed drink
  2. 3+ ingredients, at least one alcoholic and non-alcoholic each
  3. either takes some particular effort to make to make or requires exotic
     ingredients
  4. requires exotic ingredients and take some effort to make

I also slightly lowered the strenght of the drunk moodlet to encourage using
mixed drinks, though getting drunk out of regular drinks is still quite
effective due to how long lasting it is.