Enhance the power of the Shuttle Manipulator (#39870)
The 50s/5s thing appears to be a mistake, given that 5s is 50ds and fast travel checks seconds against 50 and sets deciseconds to 50 on adjacent lines.

In addition to sending shuttles to docks, three extra modes are added:

    Infinite Transit puts the shuttle in transit which lasts until another destination is chosen
    Delete Shuttle does what it sounds like
    Into The Sunset marks every mind on the shuttle as 'escaped' and then deletes the shuttle
        Also puts the mobs in stasis so they won't suffocate, bleed out, etc. before roundend

It is also now possible to change the destination dock of the arrivals shuttle (example: an event where arrivals are sent to the Lavaland Wastes dock instead).