Overhaul fire extinguishers to not use spawn/sleep (#39860)
Removing all spawns and sleeps in order to make this world a better place while
trying to keep the functionality as close as possible to the real thing.

Changed the whole code for fire extinguisher projectiles etc to not use spawn
and sleep and reduce the number of loops. Tested on private server and working.

Performance improvement, hopefully. The old code seemed pretty inefficient to
me.