Fix explosions damaging rod form wizards and supply pod contents (#40758)
The issue was that when casting rod form, the wizard is forceMoved() into the
rod. When an epicenter explosion happens, explosion.dm forcibly goes through
the contents of EVERYTHING, and calls ex_act().

This caused the wizard in the contents of the rod to get damaged by epicenter
explosions (ie if they hit a fuel tank in rod form)