[ready] Prevents trying to use explosive lances as an ingredient in crafting explosive lances (#41223)
When making an e-lance, the recipe would try to use e-lances as an ingredient, only failing after the do_after() proc. This makes it so explosive lances aren't even considered as a possible ingredient for the recipe.

cl MrDoomBringer
fix: You will no longer attempt to use an explosive lance to craft an explosive lance
/cl

This PR also lays a foundation for blacklisting other subtypes in crafting recipes.