Alien infestation event not limited to once per round, but will not fire while existing aliens are alive (#41719)
Alien infestations are not limited to once per round only, but will never
happen while existing aliens are alive.

Essentially, it's really disheartening when a larva is spawned, and it dies
quickly because the player makes a mistake or is unlucky. So, aliens MAY happen
again, but it's still a weight 5 event, so don't get your hopes up.