Makes the tooltips less annoying in the Centcom Pod launcher verb (#41821)
cl MrDoomBringer
admin: The Centcom Pod Launcher has been reorganized, making it easier to navigate at the expense of compact-ness.
/cl

Each button in the podlauncher verb had a tooltip explaining what it did, however these would get in the way of selecting things, which made life a pain. This pushes all the buttons to the right, so the tooltips should now stay out of the way of selecting buttons.