Non-emagged shuttles no longer throw atoms around (#41999)
- Normal shuttles that are not emagged will no longer throw things around.
  People will still be knocked down.
- Emagged shuttles and poorly designed shuttles will also throw things
  around on tables, as well as closets.

Shuttles were not mapped with the idea that their contents would be thrown
around violently every time they moved. This means that things like wheely
chairs, potted plants get chucked about the place. To get around this, the
"magnetic_catch" component was added, to partially bypass this effect in the
first place. But this component is only used in one place, in shuttle loading,
which means if a shuttle is custom built, suddenly and unexpectedly,
everything flies everywhere.

Instead, violent shuttle throwing should be limited to crappy shuttles,
(ie. asteroid with engines) and emagged shuttles. As such, I've removed
the magnetic component from tables and racks, to make shuttle throwing
more violent when it happens.