Port vendor icon display instead of color from Yogs (#42653)
Port of yogstation13/Yogstation-TG#2549
which is a port of Citadel-Station-13/Citadel-Station-13#7442
which is a port of OracleStation/OracleStation#50

This port visualizes items in all machines handled by _vending.dm, allowing a
person to see what they're buying before buying it. This is very useful for
clothing vendors for example, and much better than the previous system that had
vending machines choose from three colors randomly.

Credits to:
  * AndrewMontagne (Oracle, original implementation)
  * FlattestGuitar (Citadel)
  * nichlas0010 (Yogstation)