BurnMix , New ghetto healing drug with risks!
Lots to go through. Will be making a detailed wiki entry with a flow chart but here is a summary.

This is a ghetto healing drug with lots of side effects depending on how its used. The drug itself reacts to two other chemicals. Epinepherine and Perfluorodecalin. The drug increases in healing potency when any (or both) of these chems are present but in return the user suffers half of the OD symptoms of those drugs + a few others depending on the combination.
The Drug has 3 different "modes" ::
::
Non OD: The safest way to consume burnmix. Healing is snail pace but you get to avoid more serious symptoms.

OD: BurnMix increases in potency after its OD point. This is the first drug that I know of that uses this original mechanic. Healing is now more potent as well as other chems in your body. Beware serious side effects will occur as well as rolling for brain damage when mixing both Epi and Perf in this mode. Buffs previous negative side effects as well as introducing more potent OD symptoms.

Dead: You get to play a RNG healing game where you have a chance to make things worse.