BurnMix , New ghetto healing drug with risks! (#6875)
* BurnMix , New ghetto healing drug with risks!

Lots to go through. Will be making a detailed wiki entry with a flow chart but here is a summary.

This is a ghetto healing drug with lots of side effects depending on how its used. The drug itself reacts to two other chemicals. Epinepherine and Perfluorodecalin. The drug increases in healing potency when any (or both) of these chems are present but in return the user suffers half of the OD symptoms of those drugs + a few others depending on the combination.
The Drug has 3 different "modes" ::
::
Non OD: The safest way to consume burnmix. Healing is snail pace but you get to avoid more serious symptoms.

OD: BurnMix increases in potency after its OD point. This is the first drug that I know of that uses this original mechanic. Healing is now more potent as well as other chems in your body. Beware serious side effects will occur as well as rolling for brain damage when mixing both Epi and Perf in this mode. Buffs previous negative side effects as well as introducing more potent OD symptoms.

Dead: You get to play a RNG healing game where you have a chance to make things worse.

* Update medicine.dm

Fix for previous commit.

* BurnMix , New ghetto healing drug with risks!

Lots to go through. Will be making a detailed wiki entry with a flow chart but here is a summary.

This is a ghetto healing drug with lots of side effects depending on how its used. The drug itself reacts to two other chemicals. Epinepherine and Perfluorodecalin. The drug increases in healing potency when any (or both) of these chems are present but in return the user suffers half of the OD symptoms of those drugs + a few others depending on the combination.
The Drug has 3 different "modes" ::
::
Non OD: The safest way to consume burnmix. Healing is snail pace but you get to avoid more serious symptoms.

OD: BurnMix increases in potency after its OD point. This is the first drug that I know of that uses this original mechanic. Healing is now more potent as well as other chems in your body. Beware serious side effects will occur as well as rolling for brain damage when mixing both Epi and Perf in this mode. Buffs previous negative side effects as well as introducing more potent OD symptoms.

Dead: You get to play a RNG healing game where you have a chance to make things worse.

* Code seems to have duplicated (migrates computers)

Fixed the double that was generated.

Migrated computers that probably broke it.

* Restoring my main branch

Restoring my main branch because I was stupid for making my original changes in main

* Migrating burnmix to its own branch

I originally created burnmix in main like an idiot.

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Good catch!

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Update code/modules/reagents/chemistry/reagents/medicine_reagents.dm

Co-Authored-By: alexkar598 <25136265+alexkar598@users.noreply.github.com>

* Alek's suggestion while keeping the weight of the pick.

* Taste Change

no longer "dixel"

* Gets rid of dead logic

Dead people can't process chems