I'm probably just retarded, adds two new songs to titlescreen.
Adds two new songs to the title screen, Leaders of the Free World by Elbow, and How the Gods kill, by Danzig. I don't believe I should have to test this as it just adds two new songs to the audio player. Now fixed to make it add to the ticker.dm instead of config.

Changelog:
🆑
:soundadd: Added two new songs to the title screen playlist.
/🆑