Adds new achieve for CEs getting 3 MW on a station powernet
A lot of shit had to happen in order to make this work.
I had to add a new specification to powernets to mark what Z-level they exist on, to ensure that I only scan powernets that're on the station Z-level.

Additionally, SSachievements now is SS_BACKGROUND instead of SS_NO_FIRE, since I didn't want to have this CE achievement bullshit happen in the middle of the very important SSmachine tick.