Adds The Winds of Fjords by minomus, to the titlescreen. (#8752)
This will add a dazzling new song, with the edges of ruggedness that are appealing in all the right ways. The Winds of Fjords, by minomus.

Listen to it here! https://www.youtube.com/watch?v=RnQofA9CNww

🆑 Xoxeyos
rscadd: Adds 'The Winds of Fjords' by minomus, to the title screen.
/🆑