Nanite Sensor Programs Extra Setting Fix
This fix finally allows someone to edit extra settings such as "sent code" on Nanite sensor programs. The coder initialized the input boxes in the UI as null and also made it so it return itself if it were null. I simply set an initial value of 0 and resolved the issue it was having.