Stuns and Stamcrit Now Radio Mute (#9566)
* Create web_trees_and_costs.dm

* Disallows radio chatter during stuns/paralysis

While stunned or in stamcrit, prevents radio usage. Meant to go hand-in-hand with the baton changes.

* master_branch_cleanup

Removing another project from mastery

* Fixed typo

Changed underscore to closed paran in line 135

* Fixed missing underscores

Fixed typos. For real this time.

* wrong var type

* Added carbon check

Didn't realize stamcrit was carbon only. My b

* Hopefully actually the right change

But for real this time (thx theos)

* Updated sloppy code

Changed if checks on a flag to just the check itself.