Fixes Stabilized Light Pink Speed
Fixes/Re-adds Stabilized Light Pink Speed.  Failure to updade the description along with correct function in our sister codebase TG lead to speculation that it was a bug until the coder who removed it stepped forward. Speed increase is equal to .54 To compare: Meth speed is 1 and DNA Vault is .74