Fixes Anti-magic sometimes not working, fixes Runic Bombs double-exploding sometimes (#10577)
* Makes anti_magic_check mostly a /mob/living proc instead of /mob/

* (big and scary) Revamps how magic projectiles deal with antimagic

Previously, every singular ../projectile/magic had an onhit override that did the same 5-ish lines of code. No more!

As well, all of it is now handled in prehit instead of on_hit, which should fix some issues regarding spells ignoring antimagic.

* Fixes Runic Bombs detonating twice when targetting carbons

AFAIK this halves the damage of this spell. Kind of a game balance decision whether to preserve the bugged double-damage or not.

* Fixes compiletime in living.dm and warning in magic.dm