removes the tinfoil hat because it's useless and bugged
i mean, who wants to use the misc tab on an autolathe when it's hacked, right? the entire tab goes fucked over when you hack it and everything is replaced with tinfoil hats. it keeps growing in the list of hats and if you change tabs it crashes the autolathe and you have to reopen the autolathe menu.