More fixes
Adds a few firealarms and fire extinguishers i missed.
Adds four windows, one in the processing roome, one in the bar, two in the vacant room. Other machinery have been moved accordingly.
The bookcase not working, replaced with a pajama wardrobe with plushies!
Opacifies the recovery room door for privacy.