Adds digitigrade-accessible footwraps
Adds infrastructure for digitigrade-usable shoes, able to easily be added to in the future. Adds 3 types of footwraps. 2 are currently active, one for common crew and one for Security, both appearing on relevant jobs at roundstart. Both can be worn by digitigrade or straight leg races, and there are flags to set later new wraps as either straight or digitigrade leg locks.