Gathering of the Old Gods Religious Sect (#11879)
* sect of the old gods

* dont build shit instantly, actually

* slight knife buff

* golem changes

allows you to make golems yourself for a steeper price. Also gives the golems telepathy, fancy large speech complete with a spooky sound, and removes their slowdown. However, they lose their voice of god. The golems were supposed to look like reskinned runic golems but something fucked up and I don't know how to fix it, so if anyone wants to figure it out i'll keep the sprites in so they dont have to waste a couple minutes redoing them.

* less favor cost for golem, shorter ritual for meat blessing

* quality of life stuff: stops automute and makes meatblessing be the first rite in the list

makes it easier to bless meat since its the first option to show up so you dont need an extra click. Also stops it from automuting since it nows displays 1 of 3 messages instead of just the same one.

* golem armor nerf, speed buff

* nerfed synthmeat, buffed other meats, new golem power

* s

trying to fix the issue but i dont know what it is, this is just a test

* fix

* changedtobrute

* fix sprite