Adds a basic framework for skills
Adds many things as a base for a WIP skill system. Right now, this contains a skillset and skill datum, a few procs to help manipulate and read skills, a new button for a skill menu, a broken skill menu tgui, and two examples for how the system may work in the future.