Doubles event time
Its too hard to click on event cancellation prompt when admin logs are spamming at the speed of sound. I have missed event cancellations because I want to cancel it, but I don't make it on time. Or that I didn't realize because I had the audacity to look away or type for 5 seconds, and then only have a 5 second window to cancel. Doubles the event cancellation timer which doesn't really affect anything on the player's side, just lets admins more easily ward off events that really don't need to be spawned.