[admin] Doubles Admin Event Cancellation Time  (#13456)
* Doubles event time 

Its too hard to click on event cancellation prompt when admin logs are spamming at the speed of sound. I have missed event cancellations because I want to cancel it, but I don't make it on time. Or that I didn't realize because I had the audacity to look away or type for 5 seconds, and then only have a 5 second window to cancel. Doubles the event cancellation timer which doesn't really affect anything on the player's side, just lets admins more easily ward off events that really don't need to be spawned.

* Update code/modules/events/_event.dm

* Update _event.dm

* Update code/modules/events/_event.dm

* made SPOOKY RED

* FORGOR

* Demote Bibby

* Bibby blights me

cuts accidental 200 seconds to seconds

* Admin Text is correct time

* Update code/modules/events/_event.dm

Co-authored-by: Lucy <lucy@absolucy.moe>

Co-authored-by: TheGamerdk <5618080+TheGamerdk@users.noreply.github.com>
Co-authored-by: adamsong <adamsong@users.noreply.github.com>
Co-authored-by: Lucy <lucy@absolucy.moe>