PR Core
Swaps crate and table in botany.
Fixes monkey pen cabinet and action figure in genetics.
Adds a window to CMO office.
Fixes lighting in:
− Service Hall
− Xenobio Desk
− Upper Hallway
− Maintenance
Adds names to the shutter buttons and shifts them a few pixels.
Swaps genetic sequencer and disk holder in botany for quality of life.
Swaps a closet and a table so monkeys dont break the glass table every shift in genetics.
Moves around the cloning apc.
Moves around the xenobio apc.