greed slot machine has a better chance of winning
i'll see if this is too powerful, but hopefully the greed demons are making people spend money to use it regardless. Anyhow, even if you do happen to win this while spending your maximum life, you aren't even guaranteed to get something good, and there is a 10% chance to flatout die, so i'd say a higher chance of winning is in order.