swamp : drowned burial grounds | + areas
These burial grounds once housed what is presumed to be a noble swordsmen and his retinue, before it got flooded that is. The water seems to have done something strange to the bones of the dead, and now many have crawled out of their coffins to protect the treasures of their noblemen. Miners can grab a wisp lantern, weak claymore, and templar armor after fighting a group of skeletons.