Refactors condiments to support yogiments likes cinnamon
The phrasing is a bit awkward but I think this is the best way of implementing this.

Note that, since we're moving away from spammy typelists and going to a singular static, this means that new condiment types with special possible_states will have those states overwritten by initialize_possible_states unless this proc is overwritten for them (as it is for packs). If you're in git blames right now trying to figure out a bug caused by that, that's why.