Fixes dangling \the and \a macros everywhere
OpenDream caught these when I attempted to compile this codebase with it. In OD, these compile-time. In BYOND, they cause dumb behaviour.

In pretty much all cases, a macro with no buddy just becomes null or, a tab character followed by "he". None of these are desirable, so we're doing a typo run to fix them.