[s] Fixes several bugs & typos picked up by OpenDream (#15622)
* Fixes dangling \the and \a macros everywhere

OpenDream caught these when I attempted to compile this codebase with it. In OD, these compile-time. In BYOND, they cause dumb behaviour.

In pretty much all cases, a macro with no buddy just becomes null or, a tab character followed by "he". None of these are desirable, so we're doing a typo run to fix them.

* Fixes potential crash concerning speedbikes

This was causing a compiletime in OpenDream, and for good reason.

icon was null or something stupid in the previous code.

* Fixes unintended class declaration with fish crates

"crate name" actually causes a brand new type here, called /datum/supply_pack/critter/carp/crate, whose name is overridden.

This is a compiletime in OD and does the above silliness under BYOND.

EDIT: Also includes a very tiny space removal for a function call someplace, don't worry about it