Make the mindflayer small and have 6 shots
with the new rays it shoot it won't be effective by itself, this make it have a small powercell so you really understand to not use it alone and also small so you can easily carry it along something else. The mindflayer is not enough by itself  but will multiply the effectiveness of any stamina damage weapon when paired with them.