Add the mindflayer design
Yes it's sec and not armory because it's worse than getting a xray gun so there would be no reasons to take it when you have armory access. Idea is for it to be a way to make stamina effective when you don't have access to lethal and can't just stamcrit the target. It's also non lethal enough to go there.