Make the pins pricier and only sec access
why the fuck was a sec cratewith armory access anyway, secoff can now get those pins for temp guns and such, the real limitation is access to the armory for them
Now once you  depleted the firings pin of the brig each new gun or mindshield upgrading will cost you 1200 credits (one laser is 666 credits). The crate is made small so it bring less money when shipped back and also because now it's only a box you buy.