Brain damage gun (#16387)
* flayer rays now only break your mind

And potentially break it A LOT

* Make the mindflayer small and have 6 shots

with the new rays it shoot it won't be effective by itself, this make it have a small powercell so you really understand to not use it alone and also small so you can easily carry it along something else. The mindflayer is not enough by itself  but will multiply the effectiveness of any stamina damage weapon when paired with them.

* Update special.dm

* normal cell

actually it's probably fine to let it have 100 stamina damage in its battery

* Add the mindflayer design

Yes it's sec and not armory because it's worse than getting a xray gun so there would be no reasons to take it when you have armory access. Idea is for it to be a way to make stamina effective when you don't have access to lethal and can't just stamcrit the target. It's also non lethal enough to go there.

* oops

nothing there to see, move along

* add the mind flayer design to basic beam weaponry

* Update special.dm

* Delete energy.dmi

* Add files via upload

* Update mindflayer.dm

* Delete projectiles.dmi

* Add files via upload

* Delete energy.dmi

* Add files via upload

* give the flayer ray it's own sprite

* Update mindflayer.dm

* alien tech instead

Kinda fit the theme, also make it immune to balance complains because it will appear only once in a blue moon.

* Delete guns_lefthand.dmi

* Delete guns_righthand.dmi

* Add files via upload

* Give the mind flayer its own inhand sprite

* CHOKE!!

* You have our latest warcrime in your hands

* Update weapon_designs.dm