Update objective.dm (#17000)
My name is not important. What is important is what I'm going to do. I just fucking hate this world, and the human worms feasting on its carcass. My whole life is just cold, bitter HATRED. And I always wanted to die violently. This is the time of vengeance, and no life is worth saving. And I will put into the grave as many as I can. It's time for me to kill. And it's time for me to die.