[TG UPDATE] 5 - POOPER (#17312)
* Popper Tooltips 2 (they don't crash this time) (#58980)

Co-authored-by: Aleksej Komarov <stylemistake@gmail.com>

* Make tooltips use one popper, fixing mount lag (#61783)

In the last episode, I fixed update lag. This is what made the R&D console laggy to scroll through.

In this episode, we fix mount lag, which is what makes the R&D console laggy to open and search through, since it has to constantly make the new tooltips.

This is done by using a trick tippy.js uses, where they create one popper instance for multiple tooltips, then just move it around. I decided to emulate this by moving the popper stuff to Tooltip itself. The Popper component still exists as some non-tooltip consumers for it exist, such as preferences menu clothing selection. This is done through the ugliest code of my fucking life, but in a way that's totally transparent to consumers.

* Revert "Make tooltips use one popper, fixing mount lag (#61783)"

This reverts commit 180770b51963768f0b4c0118bb375b2d74382293.

* Another fix for laggy tgui tooltips -- Turn off popper event listeners (+ give R&D console its tooltips again) (#61343)

Every single popper registered TWO event listeners for scrolling and resizing. This does not appear to be necessary for our cases.
The R&D console has been given back its tooltips

* Make tooltips use one popper, fixing mount lag (#61783)

In the last episode, I fixed update lag. This is what made the R&D console laggy to scroll through.

In this episode, we fix mount lag, which is what makes the R&D console laggy to open and search through, since it has to constantly make the new tooltips.

This is done by using a trick tippy.js uses, where they create one popper instance for multiple tooltips, then just move it around. I decided to emulate this by moving the popper stuff to Tooltip itself. The Popper component still exists as some non-tooltip consumers for it exist, such as preferences menu clothing selection. This is done through the ugliest code of my fucking life, but in a way that's totally transparent to consumers.

* Make tooltips use one popper, fixing mount lag (#61783)

In the last episode, I fixed update lag. This is what made the R&D console laggy to scroll through.

In this episode, we fix mount lag, which is what makes the R&D console laggy to open and search through, since it has to constantly make the new tooltips.

This is done by using a trick tippy.js uses, where they create one popper instance for multiple tooltips, then just move it around. I decided to emulate this by moving the popper stuff to Tooltip itself. The Popper component still exists as some non-tooltip consumers for it exist, such as preferences menu clothing selection. This is done through the ugliest code of my fucking life, but in a way that's totally transparent to consumers.

* Revert "Make tooltips use one popper, fixing mount lag (#61783)"

This reverts commit 7eb9067222458edce1b362c0cfadad7de846fa73.

* Revert "Make tooltips use one popper, fixing mount lag (#61783)"

This reverts commit 346128869e7d509ccbd7907912b016150d329264.

* Revert "Another fix for laggy tgui tooltips -- Turn off popper event listeners (+ give R&D console its tooltips again) (#61343)"

This reverts commit bd3f395d76b7c914ecf9f68d41d6eddeacd96f69.

* tgui: Round Gauge (#55230)

This PR introduces the wacky round gauge for showing all of your favourite metrics in half-circle format. Show off those wacky numbers, use some scary blinking lights, feel alive!

I've also gone ahead and included this in the canister and tank (think internals) UIs. I've also done some refactoring of data sending from canisters because GOSH DANG it required some.

* GAWD DAMN IT

* EE

* Fixes Production issue

* Components

* e

* e

* Revert "e"

This reverts commit e9e804ba774a61a9a2e0c475472f8e0591b2625b.

* TOO SOON

* oops

* e

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: Aleksej Komarov <stylemistake@gmail.com>
Co-authored-by: Bobbahbrown <bobbahbrown@gmail.com>