Adds a new syndicate space ruin + new spawners (#18498)
* Adds the station as a whole, coded new areas

* Forgot to save the file

* Updated station with areas

* Updated map with essentials

* Fixed solar arrays starting unpowered

* Finished station layout, fixed some bugs

* Added the spawners code, fixed some station bugs.

* Fixed sleeper not rendering icon properly

* Addons to the map, added the spawners in.

* Made the station an actual derelict now

* Made the station an actual derelict x2

* Added lore notes to guide the spawners.

* paper

* Lore papers added to the ruin (at last)

* Made the station more derelict, added hostile mobs

* More immersion details added

* Removed soldiers, added few pirates and a gun

* More pirates, less mats, restructured map

* Converted map to TGM (second attempt)

* Fixed pistol mag incompatibility, added 1 solar assembly

Fixed the incompatibility of pistol mags, now the mag should be able to be used with the gun. Added 1 more solar assembly to match the slots for solars.

* Added protolathe and printer boards

Added the missing protolathe and circuit imprinter boards.

* Scattered plastitanium has more sheets

Added more sheets to the scattered plastitanium and plastitanium glass.

* Added extra mag, added esword crate

Added another mag for the pistol (for the weaklings). Added a secure crate with 2 eswords in the virology lab.

* Minor fixes - Hydroponics APC fixed (hopefully)

Changed a few area placements, fixed (hopefully) the hydroponics APC position.

* Added more ores (and missing ones)

The ore crate spawned lacking some ores such as uranium and dilithium, now the crate spawns with those. Added more bluespace as well.

* Added disposal system

Disposal system tubes added to the ruin. They're unfinished, same as the atmos pipes. Added missing disposal bins to some rooms.

* big red text update

Updates the red text to specify that if you touch the main station or abandon the derelict to fuck off somewhere else (without intention to return) you shall no longer be welcome to my ruin.