Sweeping Melee Attacks (#21025)
This adds the ability for some weapons to attack in an arc instead of a precise hit, when clicking on a non-adjacent tile on HARM intent, and wielded if the weapon is two-handed. Sweeping attacks have a 50% longer attack cooldown than precise hits, and blunt weapons can't hit more than one target at once.