Bolton's Assortment of Costumes! (#3009)
* This is my repository right?

If not then I'm surprised

* Slight Fix

Just remembered I forgot to finish a description, a bit embarrassing! Now it's completed.

* Changed to Secdrobe

Made more sense to change it to the Secdrobe instead of the autodrobe, as it was a cop mascot costume in its description, it doesn't make sense for it to not be used by security, IMO.

* Tweaking

Tweaked some more elements of the descriptions and design of the suit.

* Added More Costumes!

The original was pretty small, so adding this gives it a lot more justification, plus adds in a few things we didn't have, like a module for glasses, which previously was missing due to people improperly placing items into the TG code (I think.)

* Tweak to CS Glasses

Accidentally had the code a little bit off, with the dmi file mislabeled. Fixed now!

* Added the metrocop uniform

Here is the uniform that the metrocops use in Half Life 2, added to my costumes and stuff.

* Citizen Uniform

To match the metrocop!