Some minor runtime fixes (#3130)
* Fix a bunch of various runtimes

* fix prefs bullshittery

* fix that runtime

* Fix balloon alert runtimes when draining a heretic influence (#85560)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

While this runtime log is from MonkeStation, the associated code is the
same on /tg/, and thus the bugfix applies to both

```
[2024-08-02 22:29:45.372] runtime error: addtimer called with a callback assigned to a qdeleted object. In the future such timers will not be supported and may refuse to run or run with a 0 wait
 - proc name:  stack trace (/proc/_stack_trace)
 -   source file: code/__HELPERS/stack_trace.dm,4
 -   usr: Macie Carmichael (/mob/living/carbon/human)
 -   src: null
 -   usr.loc: the floor (102,80,2) (/turf/open/floor/iron/dark)
 -   call stack:
 -  stack trace("addtimer called with a callbac...", "code/controllers/subsystem/tim...", 621)
 -  addtimer(/datum/callback (/datum/callback), 12.45, 0, null, "code/modules/balloon_alert/bal...", 89)
 - the rising blemish (/obj/effect/heretic_influence): balloon alert perform(Macie Carmichael (/mob/living/carbon/human), "influence drained")
 - the rising blemish (/obj/effect/heretic_influence): balloon alert(Macie Carmichael (/mob/living/carbon/human), "influence drained")
```

## Why It's Good For The Game

bugfix, nya~

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Fixed a runtime error after draining a heretic influence.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* fix gravitum runtimes (gravgen did this too so its prolly fine and not a janky hack)

* rfijrfjunrefujhnerjnhr

* fix wound bs

* yeah guess we're doing this now

* `QDEL_LAZYASSOCLIST`

* fix organ fab runtime

* that is not a list!!!

* fix null weight

* fix some ez nanite runtimes

* fix shadow runtime

* fix jukebox screwdriver

* chatmessage runtimes

* no more runtime if you get nuked or whatever lol

* me when slimes

* fix ownerless antag datum prototypes being stored alongside actual antags