playsound() will runtime if you try to pass a list into it. (#86934) (#3742)
## About The Pull Request

Adds an islist() sanity check and runtime to playsound() to prevent
client crashing.
## Why It's Good For The Game

Clients have a possibility to crash if you pass a list() into
playsound() without runtimming. It's sporadic. It's not consistant. I've
not managed to reproduce it on local. But monkey ran into this issue
before and I've ran into this issue if you forget to use pick()
 
Plus there should honestly just be something to tell people to not pass
lists into something that plays a single sound.

## Changelog

Not player facing

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>