Various DB-related ports from /tg/ (#4028)
* Stop stack tracing if ezdb daemon is already running (#76555)

## About The Pull Request

This is an annoyance for me because I often kill the game but not the
EZDB daemon (leaving the cmd window open), and it just triggers this
stack trace while I have a debugger running. It's still able to connect
to the daemon in-game and functions perfectly fine, just hate dealing
with a completely needless stack_trace() every single time.


![image](https://github.com/tgstation/tgstation/assets/34697715/f8ae3cec-1cf6-4e70-b9bc-c7088121b85c)


![image](https://github.com/tgstation/tgstation/assets/34697715/a13774b8-addf-43fd-82f0-3d0e644d2dc1)
## Why It's Good For The Game

LET ME WORK!!! PLEASE!!!
## Changelog
Nothing that players should care about.

* Wait for db queries to finish during db shutdown (#77311)

* Remove delayed in MassInsert (#80196)

Necessary to update MariaDB to a newer version. Was necessary until we
had async queries, apparently.

* Fix shutdown hanging if the db went away. gives shutdowns a time out. improves db shutdown logging (#81813)

terry right now:
```
[20:11:31] Runtime in code/controllers/subsystem/dbcore.dm,353: Attempting to create a new db query during the world shutdown
...
[21:25:29] Runtime in code/controllers/subsystem/dbcore.dm,353: Attempting to create a new db query during the world shutdown
```

* refactor db connection timeouts (#81816)

Give it exponential back off and smerter logic.

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: oranges <email@oranges.net.nz>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>