Misc quirk related improvements (#4013)
* Adds special lung sprites for smoker quirk (#78119)

What it says on the tin, adds special sprites for ethereal, plasmaman,
and regular lungs for if you use the smoker quirk.

Top is normal lungs, bottom is lungs if you use the smoker quirk

![smokerlungs](https://github.com/tgstation/tgstation/assets/58376695/eeb95ab6-3f99-4f64-8cfb-6d9ded0b9413)

Adds some cool flavor and a visual difference between regular and smoker
lungs. Previously, smoker lungs looked identical to normal ones but had
less health and healed slower.

Details are nice, and it was weird there was no visual difference before
since realistically, smoking constantly makes your lungs very visually
different and deteriorate over time.

:cl:
image: adds special sprites for lungs when you use the smoker quirk
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>

* Funny clown internals (#77963)

This PR changes the internals that spawn inside the clown's survival box
for a new one with a rainbow sprite, higher O2 volume (same as the engi
ones) and a secondary gas on top of O2 to make things more interesting
for the clowns.
The gas options are:
BZ, which just adds hallucinations for the clown, without the brain
damage effect as it is in low percentages.
N2O will make the clown giggle and laugh, without the sleep.
Helium will give the clown a "funny voice".

These tanks are also added to the mail list of clown fans and the clown
costume crate at cargo.

And codersprites, I can polish them later if people think it is pixel
soup, I'm not happy with them that much, but making this looks good
might be above my paygrade...
<details><summary>Pics here</summary>
<p>

![clown_internals](https://github.com/tgstation/tgstation/assets/55374212/f5eda877-a01a-4dfa-b481-7d406c4fb768)

![in game clown
internals](https://github.com/tgstation/tgstation/assets/55374212/342285ae-919b-49ab-a97e-cdf25a975f83)

</p>
</details>

The main goal I have with this is to add more uses for Atmos Content to
other players in a flavorful way.
Atmos is not something the crew interacts in a positive way often and I
want to change that.

These tanks are something quite minor but flavorful IMO, also will make
people know Helium fucking exists...

The tanks *shouldn't* change much of the clown's round in a negative
way, and the default O2 internals are in every hallway's locker so even
if they don't want to deal with the hallucinations it is not a big deal
to dodge them.

:cl: Guillaume Prata
add: New funny internals for the clowns to spawn with. They come with O2
and a secondary gas between 3 options: BZ, Helium and N2O. Talk with a
"different tone" with Helium, giggle and laugh "uncontrollably" while
under the minor effects of N2O or have "fun" hallucinations while under
the minor effects of BZ.
balance: To not cut on how long the clown's O2 internals last due to the
mixed gases, the funny internals have 50% more gas volume, same as
engineers' internals.
/:cl:

---------

Co-authored-by: CRITAWAKETS <sebastienracicot@hotmail.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>

* Improve the new clown internals

* move

* Breaks up the three quirk code files into individual quirks (#78407)

Separates all the quirts out of the positive, neutral and negative files
into their own files. The smoker and alcoholic quirks are left with
junkie as they are subtypes of it.

These three code files were chonkers at 400, 483 and 1564 lines. They
were run on files of types only connected by all being quirks, and made
it hard to navigate these files without the user's eye sliding from one
quirk to the next.

Breaking them down makes it easy to look at the code for a specific
quirk indepentently, and also allows looking between multiple quirks
without having to scroll around the same file.

No player facing changes

* You can now pick the limb you want replaced with the Prosthetic Limb Quirk. (#78517)

## About The Pull Request

The Prosthetic Limb quirk allows you to pick the limb you want. The
default is still random
## Why It's Good For The Game

Good for character customization, which is always nice. Statics who got
a specific limb replaced can now represent that properly in-game.
Also, this makes the quirk show up on the dummy on the character
customization again. This *does* cause a slight issue where when you
switch to another character, it keeps the previous' skin color, but it's
minor enough that i think the benefits are worth the drawback. If asked
i will make it not appear on the dummy though.
## Changelog
DATA_, with great help from Kapu:cl:
add: You can now choose the prosthetic you want with the Prosthetic Limb
quirk.
/:cl:

* Enable pride pins

* cleanup disabled quirks

* Apply fixes for integration tests

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: GuillaumePrata <55374212+GuillaumePrata@users.noreply.github.com>
Co-authored-by: CRITAWAKETS <sebastienracicot@hotmail.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: Thunder12345 <Thunder12345@users.noreply.github.com>
Co-authored-by: DATAxPUNGED <44149906+DATA-xPUNGED@users.noreply.github.com>