Hotfix chemmaster (#4523)
* Update chem_master.dm

* Adds checks for plumbing pill press (#88002)

- Fixes #88001. Pllumbing pill press now reads the maximum volume of the
selected product instead of showing a constant of 50u. Patches have a
maximum volume of 40u so we send that to the UI
- Plumbing pill press now validates the selected container to see if
it's a valid printable option to prevent href exploits. It has the same
issue as #87779 but now that's fixed
- Plumbing pill press checks to see if the input volume is a number and
returns false to stop the UI from updating if not

:cl:
fix: plumbing pill press advertises the correct maximum volume of your
selected product instead of always 50u on the UI
code: plumbing pill press validates selected container to prevent href
exploits
code: plumbing pill press validates input volume to see if it's a number
/:cl:

* Update pill_press.dm

* Update pill_press.dm

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>