[PORT] Heretics dream of objects around unused influences (#4970)
* Light dreaming refactor + new dream (#78996)

Makes it so new dream types can be more easily added to the game with
effects within them beyond just some text being displayed if you want. I
added a dream that plays a random sound file available in the game files
to demonstrate how it's used. Mainly I wanted this out of the way for
another thing I'm working on.

I also made it so if the sound subsystem fails to reserve a channel it
throws a runtime since this happening is likely always going to result
in a bug, like it did while I was testing this.

:cl:
add: New dream that plays sound at you
/:cl:

* Changes `dreaming` variable on carbons to a trait (#79940)

The documentation lied, it was a boolean that we only used in one file,
so lets just make it a trait and cut down on how much scrolling the
average person has to do when looking at carbon variables
Does not concern players.

* Heretics dream of objects around unused influences (#89003)

![image](https://github.com/user-attachments/assets/5fee2111-289b-46c6-82f8-096fcfadc50f)

![image](https://github.com/user-attachments/assets/1402896d-4d39-4168-a105-7d28bd199999)

![image](https://github.com/user-attachments/assets/20a7a1fd-3f63-4657-9834-0f568f241490)

Heretics can now traverse Mansus in their dreams to see into the real
world through the perspective of unused essences, if any.

They get a mood debuff for five minutes while doing so, and dream
regularly until recovered.

Essences sometimes spawn in closed locations (like vault, telecomms,
toilet room) and stay untapped till the end of the round.

This change should promote a more covert way of gaining a roundstart
knowledge boost for heretics.

They may dream of completely random and useless objects, but it will
still help them to know that there is still an unused essence somewhere.

:cl:
add: Heretics can now get clues about objects around unused influences
in their dreams, with a 5 minute cooldown
/:cl:

* whoops remove that

---------

Co-authored-by: Emmett Gaines <ninjanomnom@gmail.com>
Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Andrew <mt.forspam@gmail.com>